fixed int cast

+++ src/main/java/ubc/cosc322/COSC322Test.java
@@ -28,7 +28,7 @@ public class COSC322Test extends GamePlayer {
      * @param args for name and passwd (current, any string would work)
      */
     public static void main(String[] args) {
-        COSC322Test player = new COSC322Test("Player" + Math.random() * 10000, "1");
+        COSC322Test player = new COSC322Test("Player" + (int)Math.random() * 10000, "1");
 
         if (player.getGameGUI() == null) {
             player.Go();
